fix: 디렉토리 분리, 수정
