+++ src/main/java/project/linkarchive/backend/relationship/controller/RelationshipQueryController.java
@@ -1,0 +1,4 @@
+package project.linkarchive.backend.relationship.controller;
+
+public class RelationshipQueryController {
+}
