+++ src/main/java/project/linkarchive/backend/relationship/service/RelationshipQueryService.java
@@ -1,0 +1,4 @@
+package project.linkarchive.backend.relationship.service;
+
+public class RelationshipQueryService {
+}
